ソースファイル構成を見直した
