+++ zlearning/src/lib.rs
@@ -4,6 +4,7 @@
 
    hidekuno@gmail.com
 */
+pub mod param;
 pub mod tree;
 pub mod visitor;
 pub mod walker;
